requirements.txt
Django
MySQL-python
Pillow
django-crispy-forms==1.6.1
python-dateutil